porting to python 3

+++ testproject/store/views.py
@@ -7,7 +7,7 @@ from store.models import Product, ProductRating
 def rate_product(request, product_id):
     
     if request.method == 'POST':
-        product = get_object_or_404(Product, pk = long(product_id))
+        product = get_object_or_404(Product, pk = int(product_id))
         product.productrating_set.create(rating = int(request.POST.get('rating')))
         return HttpResponse(status = 200)
     else:
